requirements.txt
Django==3.1.3
mysqlclient==2.0.3
djangorestframework==3.12.2
django-filter==2.4.0
django-notifications-hq==1.6.0
django-storages==1.10.1
boto3==1.16.36
python-memcached==1.59
redis==3.5.3
python-dateutil==2.8.1
celery==5.1.2
django-ratelimit==3.0.1